Implement pre-defining structs/unions
When a struct or a union is already defined in a lib with the same
name as the transformer would generated, use that defined structure
instead of generating a new one. One particular usecase for this is
defining structures that use bitfields manually and then using those
instead of the incorrectly generated by crystal_lib.